requirements.txt
requests
Pillow
python-dotenv
openai
langchain
Flask
